Implement common base class for CirqSimulator and QSimSimulator
